refactor(inspector): move inspector-host-types into the overlay module
The host-link types live next to the host bridge that uses them, instead of
under dev-tools. The devtools-protocol bridge itself is untouched here; its
deletion is split into a follow-up.
