+++ packages/jazz-tools/src/dev-tools/index.ts
@@ -14,12 +14,3 @@ export {
   type DevtoolsBridgeConnector,
 } from "./extension-panel.js";
 export { createParentWindowBridgePort } from "./parent-window-port.js";
-export {
-  INSPECTOR_HOST_GLOBAL,
-  INSPECTOR_SUBSCRIPTIONS_MESSAGE,
-  serializeActiveSubscriptions,
-  type InspectorConnectionConfig,
-  type InspectorSubscription,
-  type InspectorSubscriptionsMessage,
-  type JazzInspectorHost,
-} from "./inspector-host-types.js";
